WIP: Auth for cross-contract calls + LP reserve property
There is a very interesting reason the following property fails to
hold:

```
invariant [reserve_a_equals_balance]
    (l.a.b.balance.pre(l._this) & l.a.b.balance.map(l._this, l.reserve_a.v))
    | (~l.a.b.balance.pre(l._this) & l.reserve_a.v = integer.zero)
```

The reason is the withdraw() transition in `liquidity_contract.ivy`.
More precisely, there is nothing (as of before this commit) that
prevents the LP contract itself from owning shares in the LP and
withdrawing them. This is not actually possible in the code, but
there is no invariant to state this as of now.

The trouble this creates is that in this situation, the reserve
field is decreased, but the balance of tokens the LP contract holds
doesn't in fact decrease -- the tokens are transferred to the LP
contract itself upon withdrawal, which is a no-op.

I tried to solve this issue by more faithfully modelling the Soroban
auth logic, to require that contracts that make a call provide
authorization before the call is made. (This is done implicitly by the
Soroban environment.)

That lead to the uncovering of another missing invariant clause,
namely that the contract addresses in the (Ivy) environment are
reflected as such in the modelled Soroban environment, i.e.
in `env.contract_addresses`.

The issue I'm running in now is that there seems to be no way to
specify the order in which `after init` blocks run. I was expecting
them to run in `include` order, i.e. the deepest runs first, but
that doesn't seem to be the case.

The concrete problem I'm seeing is that all the `_register_contract`
stuff we have to keep track of contract addresses gets clobbered by
`soroban.ivy`'s `environment` `after_init`.