WIP: something wrong with Ivy->mypyvy translation
Ivy says `suzuki_kasami.ivy` is OK, whereas the translated version
does not pass mypyvy invariant checks. I've manually edited the
translated version to make checking the correspondence to the original
easier, but I can't find the issue.

Look at the `init` and `ext_c_rcv_request` in `suzuki_kasami_manual.pyv`
and compare with the equivalent in `suzuki_kasami.ivy`. What's wrong?

I recall that in the past when I've tried mypyvy, I've always had the
impression if-then-else didn't behave as I expected, so I always had
to remove branching from my specs to make them "work". Could it be that
there is some if-then-else bug in mypyvy?