WIP: prove `both_zero_or_none_zero`
It turns out that this (despite its apparent simplicity) is an
interesting property to prove. Recall that we want to prove it because
it is necessary to ensure that deposits do not panic.

First of all, it is NOT sufficient to enforce this in `deposit` itself
(by checking that both deposited amounts are strictly positive). We
must also enforce it in `swap`, by ensuring that the new reserve amounts
are strictly positive. (Reserve amounts can be changed in `swap` as
well, not only in `deposit`.)

This commit adds that check to `swap`. It also adds a check that the `to`
address in `swap` is not the pool address, which we've forgotten previously.

It also makes progress towards proving the property by adding supporting
invariant clauses:

```
invariant [total_shares_gte_zero]
  integer.gte(total_shares, integer.zero)

invariant [total_shares_gte_any_share_balance]
  forall A, V. token_s.balance.map(A, V) -> integer.gte(total_shares, V)
```

- `total_shares_gte_zero` replaces `reserves_empty_iff_no_shares`, which
was not actually true.

- `total_shares_gte_any_share_balance` eliminates a CTI with an
  unreachable pre-state, where the pool has more shares than
`total_shares`; this cannot happen

`total_shares_gte_any_share_balance` is not preserved by `deposit`,
however. Upon inspection of the CTI, it turns out this is due to the
inadequacy of our properties for integers. (The property is indeed
preserved by actual integers.) `plus_eq_results` was added to progress
towards eliminating this spurious CTI -- this is still not done.

Before focusing again on enhancing our decidable model of integers, I
will try to establish `both_zero_or_none_zero` (and the supporting
`total_shares_gte_any_share_balance`) with actual integers for the
remaining `withdraw` transition.