Seemingly reaching the limits of decidable ints
The counter-example I get now to `total_shares_gte_any_share_balance` in
`deposit` seems to not be excludable whilst remaining in the decidable
fragment.

I get something like:

```
integer.zero = 0
integer.one = 4

0 < 4 < 3 < 2 < 1

(< is the lt relation)
```

but with no way to produce 3 from 4 -- in particular, 4 is not the
integer two.